first pass at detailing the methods
I tried to improve the description of the GLS models and be more specific about which responses variables are analysed in the different steps.
There is still some work but I think we cam go ahead with first round of coauthor reviews.
Doing a pull request so that you can check what I changed before accepting. and in case you've made changes while I was editing. This is the the .Rmd document, I didn't knit it.